Updated to allow the specification of a different uart to SPISerial Command wifly.setUart(Stream* newUart); should be issued before the wifly.begin() if a different UART is required Several other changes - most notably the use of findInResponse in sendCommand However most callers do not check the error return on this yet.

+++ src/WiFly/SpiUart.cpp
@@ -155,7 +155,7 @@ byte SpiUartDevice::readRegister(byte registerAddress) {
 }
 
 
-byte SpiUartDevice::available() {
+int SpiUartDevice::available() {
   /*
   
     Get the number of bytes (characters) available for reading.
